Add prune function to avoid leaking proc objects.
A function could start lots of processes with subprocess.popen but not
call a wait function. This would leave the table of processes full of
objects (in order for subprocess.wait to work). These would otherwise
never get garbage-collected.

The prune function polls each proc in the table. If a proc has finished,
it is removed from the table (in do_waitpid) and can then be
garbage-collected.